IDI-141: Initial upgrade to WordPress 3.9.2

+++ wp-content/plugins/role-scoper/admin/admin_lib_rs.php
@@ -20,7 +20,6 @@ add_filter('wp_dropdown_pages', array('ScoperAdminLib', 'flt_dropdown_pages') );
 if ( strpos( $_SERVER['REQUEST_URI'], 'nggallery' ) ) // Role Scoping for NGG calls ScoperAdminUI::dropdown_pages
 	require_once( dirname(__FILE__).'/admin_ui_lib_rs.php' );
 
-
 class ScoperAdminLib {
 	// filter page dropdown contents for Page Parent controls; leave others alone
 	function flt_dropdown_pages($orig_options_html) {
@@ -462,8 +461,14 @@ class ScoperAdminLib {
 			
 		$user_clause = "user_id IN ('" . implode("', '", $user_ids) . "')";
 		
-		$table_name = ( $blog_id_arg ) ? $wpdb->base_prefix . $blog_id_arg . '_' . 'user2role2object_rs' : $wpdb->user2role2object_rs;
+		if ( $blog_id_arg )
+			$table_name = ( $blog_id_arg > 1 ) ? $wpdb->base_prefix . $blog_id_arg . '_' . 'user2role2object_rs' : $wpdb->base_prefix . 'user2role2object_rs';
+		else
+			$table_name = $wpdb->user2role2object_rs;
 		
+		if ( ! $wpdb->get_results( "SHOW TABLES LIKE '$table_name'" ) )
+			return;
+
 		scoper_query("DELETE FROM $table_name WHERE $user_clause");
 		
 		foreach ( $user_ids as $user_id ) {
@@ -624,7 +629,7 @@ class ScoperAdminLib {
 		if ( ! is_array( $dismissals ) )
 			$dismissals = array();
 
-		$msg_id = ( isset( $_REQUEST['msg_id'] ) ) ? $_REQUEST['msg_id'] : 'pp_offer';
+		$msg_id = ( isset( $_REQUEST['msg_id'] ) ) ? $_REQUEST['msg_id'] : 'ppcore_offer';
 		$dismissals[$msg_id] = true;
 		update_option( 'scoper_dismissals', $dismissals );
 	}
